Added changelog with foreign keys. Minor changes to changelog-0.1
