+++ docker-compose.yml
@@ -6,7 +6,7 @@ version: '2'
 
 services:
   postgres:
-    image: postgres:10
+    image: postgres:13
     ports:
       - 5432:5432
     volumes:
