+++ src/main/java/ru/ekaripov/contactsdb/model/Photo.java
@@ -1,4 +1,4 @@
-package ru.ekaripov.contactsdb.entities;
+package ru.ekaripov.contactsdb.model;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
