+++ src/main/resources/db/changelog/changelog-0.4.yaml
@@ -1,0 +1,110 @@
+#foreign keys for tables
+databaseChangeLog:
+  - changeSet:
+      id: 4
+      author: ekaripov
+      changes:
+        - addForeignKeyConstraint:
+            constraintName: fk_authorities_users
+            baseTableName: authorities
+            baseColumnNames: user_id
+            referencedTableName: users
+            referencedColumnNames: id
+            onDelete: CASCADE
+            onUpdate: CASCADE
+        - addForeignKeyConstraint:
+            constraintName: fk_photo_person
+            baseTableName: person
+            baseColumnNames: photo_id
+            referencedTableName: photo
+            referencedColumnNames: id
+            onDelete: SET NULL
+            onUpdate: CASCADE
+        - addForeignKeyConstraint:
+            constraintName: fk_category_person
+            baseTableName: person
+            baseColumnNames: category_id
+            referencedTableName: person_category
+            referencedColumnNames: id
+            onDelete: CASCADE
+            onUpdate: CASCADE
+        - addForeignKeyConstraint:
+            constraintName: fk_contact_type
+            baseTableName: contacts
+            baseColumnNames: type_id
+            referencedTableName: contact_type
+            referencedColumnNames: id
+            onDelete: CASCADE
+            onUpdate: CASCADE
+        - addForeignKeyConstraint:
+            constraintName: fk_person_contacts
+            baseTableName: contacts
+            baseColumnNames: person_id
+            referencedTableName: person
+            referencedColumnNames: id
+            onDelete: CASCADE
+            onUpdate: CASCADE
+        - addForeignKeyConstraint:
+            constraintName: fk_history_event_types
+            baseTableName: event_types
+            baseColumnNames: history_type_id
+            referencedTableName: history_type
+            referencedColumnNames: id
+            onDelete: SET NULL
+            onUpdate: CASCADE
+        - addForeignKeyConstraint:
+            constraintName: fk_event_type
+            baseTableName: events
+            baseColumnNames: type_id
+            referencedTableName: event_types
+            referencedColumnNames: id
+            onDelete: CASCADE
+            onUpdate: CASCADE
+        - addForeignKeyConstraint:
+            constraintName: fk_events_users
+            baseTableName: events
+            baseColumnNames: user_id
+            referencedTableName: users
+            referencedColumnNames: id
+            onDelete: CASCADE
+            onUpdate: CASCADE
+        - addForeignKeyConstraint:
+            constraintName: fk_events_person
+            baseTableName: events
+            baseColumnNames: person_id
+            referencedTableName: events
+            referencedColumnNames: id
+            onDelete: CASCADE
+            onUpdate: CASCADE
+        - addForeignKeyConstraint:
+            constraintName: fk_history_type
+            baseTableName: history
+            baseColumnNames: type_id
+            referencedTableName: history_type
+            referencedColumnNames: id
+            onDelete: CASCADE
+            onUpdate: CASCADE
+        - addForeignKeyConstraint:
+            constraintName: fk_history_person
+            baseTableName: history
+            baseColumnNames: person_id
+            referencedTableName: person
+            referencedColumnNames: id
+            onDelete: CASCADE
+            onUpdate: CASCADE
+        - addForeignKeyConstraint:
+            constraintName: fk_history_users
+            baseTableName: history
+            baseColumnNames: user_id
+            referencedTableName: users
+            referencedColumnNames: id
+            onDelete: CASCADE
+            onUpdate: CASCADE
+        - addForeignKeyConstraint:
+            constraintName: fk_person_tag
+            baseTableName: tag
+            baseColumnNames: person_id
+            referencedTableName: person
+            referencedColumnNames: id
+            onDelete: CASCADE
+            onUpdate: CASCADE
