+++ src/main/resources/db/changelog/changelog-master.yaml
@@ -5,5 +5,7 @@ databaseChangeLog:
       file: src/main/resources/db/changelog/changelog-0.2.yaml
   - include:
       file: src/main/resources/db/changelog/changelog-0.3.yaml
+  - include:
+      file: src/main/resources/db/changelog/changelog-0.4.yaml
 #  - include:
 #      file: classpath:db/changelog/changelog-1.1.yaml
